♻️ Refactor API endpoints
