+++ server/blueprints/api/subsonic/search.py
